Create sitemap.php
This code does the following:

Registers a new rewrite rule for the sitemap.xml file, so that it can be accessed at a custom URL (e.g. http://example.com/sitemap.xml).

Adds a custom query variable called "sitemap" that is used to trigger the generation of the sitemap.

Generates the sitemap and outputs it as an XML file when the "sitemap" query variable is set.

Pings search engines like Google and Bing when a post or page is published or updated, to let them know that the sitemap has been updated.

Adds a link to the sitemap in the footer of the page.