Update sitemap.php
This code does the following:

Defines a constant called SITEMAP_PLUGIN_OPTIONS that is used to store the plugin's options in the database.

Registers a new rewrite rule for the sitemap.xml file, so that it can be accessed at a custom URL (e.g. http://example.com/sitemap.xml).

Adds a custom query variable called "sitemap" that is used to trigger the generation of the sitemap.

Generates the sitemap and outputs it as an XML file when the "sitemap" query variable is set.

Pings search engines like Google and Bing when a post or page is published or updated, to let them know that the sitemap has been updated. This behavior can be turned on or off using the settings page.

Adds a link to the settings page in the plugin list.

Registers the settings page, which allows the user to turn on or off the search engine pinging and footer link features.

Saves the plugin's options to the database when the settings page is submitted.

Flushes the rewrite rules when the plugin is activated or deactivated, to ensure that the sitemap.xml file is accessible.